# server/docs/swagger.yaml
basePath: /
definitions:
  api.Pagination:
    properties:
      all:
        type: boolean
      index:
        type: integer
      limit:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  incident.AddConference:
    properties:
      type:
        $ref: '#/definitions/incident.ConferenceType'
    type: object
  incident.Attachment:
    properties:
      description:
        type: string
      type:
        $ref: '#/definitions/incident.AttachmentType'
      url:
        type: string
    type: object
  incident.AttachmentType:
    enum:
    - link
    type: string
    x-enum-varnames:
    - Link
  incident.Channel:
    properties:
      id:
        type: string
      name:
        type: string
      source:
        $ref: '#/definitions/incident.ChannelSource'
      status:
        $ref: '#/definitions/incident.ChannelStatus'
      url:
        type: string
    type: object
  incident.ChannelSource:
    enum:
    - slack
    type: string
    x-enum-varnames:
    - SlackSource
  incident.ChannelStatus:
    enum:
    - operational
    type: string
    x-enum-varnames:
    - Operational
  incident.Conference:
    properties:
      conferenceId:
        type: string
      type:
        $ref: '#/definitions/incident.ConferenceType'
      url:
        type: string
    type: object
  incident.ConferenceType:
    enum:
    - zoom
    type: string
    x-enum-varnames:
    - Zoom
  incident.CreateRequest:
    properties:
      addConference:
        $ref: '#/definitions/incident.AddConference'
      attachments:
        items:
          $ref: '#/definitions/incident.Attachment'
        type: array
      channels:
        items:
          $ref: '#/definitions/incident.Channel'
        type: array
      description:
        type: string
      environments:
        items:
          $ref: '#/definitions/incident.Environment'
        type: array
      functionalities:
        items:
          $ref: '#/definitions/incident.Functionality'
        type: array
      identifier:
        type: string
      incidentChannel:
        $ref: '#/definitions/incident.IncidentChannel'
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/incident.Service'
        type: array
      severity:
        $ref: '#/definitions/incident.Severity'
      status:
        $ref: '#/definitions/incident.Status'
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/incident.Type'
    required:
    - identifier
    - incidentChannel
    - name
    - severity
    - summary
    - type
    type: object
  incident.CreateResponse:
    properties:
      accountIdentifier:
        type: string
      active:
        type: boolean
      attachments:
        items:
          $ref: '#/definitions/incident.Attachment'
        type: array
      channels:
        items:
          $ref: '#/definitions/incident.Channel'
        type: array
      conferenceDetails:
        items:
          $ref: '#/definitions/incident.Conference'
        type: array
      correlationID:
        type: string
      createdAt:
        type: string
      createdBy:
        $ref: '#/definitions/utils.UserDetails'
      description:
        type: string
      environments:
        items:
          $ref: '#/definitions/incident.Environment'
        type: array
      functionalities:
        items:
          $ref: '#/definitions/incident.Functionality'
        type: array
      id:
        type: string
      identifier:
        type: string
      incidentChannel:
        $ref: '#/definitions/incident.IncidentChannel'
      name:
        type: string
      orgIdentifier:
        type: string
      projectIdentifier:
        type: string
      removed:
        type: boolean
      removedAt:
        type: string
      roles:
        items:
          $ref: '#/definitions/incident.Role'
        type: array
      services:
        items:
          $ref: '#/definitions/incident.Service'
        type: array
      severity:
        $ref: '#/definitions/incident.Severity'
      stages:
        items:
          $ref: '#/definitions/incident.Stage'
        type: array
      status:
        $ref: '#/definitions/incident.Status'
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/incident.Type'
      updatedAt:
        type: string
      updatedBy:
        $ref: '#/definitions/utils.UserDetails'
    required:
    - active
    - id
    - identifier
    - name
    - severity
    - status
    - summary
    type: object
  incident.CreateResponseDTO:
    properties:
      correlationId:
        type: string
      data:
        $ref: '#/definitions/incident.CreateResponse'
      message:
        type: string
      status:
        type: string
    type: object
  incident.Environment:
    properties:
      environmentId:
        type: string
      environmentName:
        type: string
    type: object
  incident.Functionality:
    properties:
      functionalityId:
        type: string
      functionalityName:
        type: string
    type: object
  incident.Incident:
    properties:
      accountIdentifier:
        type: string
      active:
        type: boolean
      attachments:
        items:
          $ref: '#/definitions/incident.Attachment'
        type: array
      channels:
        items:
          $ref: '#/definitions/incident.Channel'
        type: array
      conferenceDetails:
        items:
          $ref: '#/definitions/incident.Conference'
        type: array
      createdAt:
        type: string
      createdBy:
        $ref: '#/definitions/utils.UserDetails'
      description:
        type: string
      environments:
        items:
          $ref: '#/definitions/incident.Environment'
        type: array
      functionalities:
        items:
          $ref: '#/definitions/incident.Functionality'
        type: array
      id:
        type: string
      identifier:
        type: string
      incidentChannel:
        $ref: '#/definitions/incident.IncidentChannel'
      name:
        type: string
      orgIdentifier:
        type: string
      projectIdentifier:
        type: string
      removed:
        type: boolean
      removedAt:
        type: string
      roles:
        items:
          $ref: '#/definitions/incident.Role'
        type: array
      services:
        items:
          $ref: '#/definitions/incident.Service'
        type: array
      severity:
        $ref: '#/definitions/incident.Severity'
      stages:
        items:
          $ref: '#/definitions/incident.Stage'
        type: array
      status:
        $ref: '#/definitions/incident.Status'
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/incident.Type'
      updatedAt:
        type: string
      updatedBy:
        $ref: '#/definitions/utils.UserDetails'
    required:
    - active
    - id
    - identifier
    - name
    - severity
    - status
    - summary
    type: object
  incident.IncidentChannel:
    properties:
      slack:
        $ref: '#/definitions/incident.Slack'
      type:
        $ref: '#/definitions/incident.IncidentChannelType'
    type: object
  incident.IncidentChannelType:
    enum:
    - slack
    type: string
    x-enum-varnames:
    - ChannelSlack
  incident.ListResponse:
    properties:
      content:
        items:
          $ref: '#/definitions/incident.Incident'
        type: array
      correlationID:
        type: string
      pagination:
        $ref: '#/definitions/api.Pagination'
    type: object
  incident.ListResponseDTO:
    properties:
      correlationId:
        type: string
      data:
        $ref: '#/definitions/incident.ListResponse'
      message:
        type: string
      status:
        type: string
    type: object
  incident.Role:
    properties:
      roleType:
        $ref: '#/definitions/incident.RoleType'
      userDetails:
        $ref: '#/definitions/utils.UserDetails'
    type: object
  incident.RoleType:
    enum:
    - incidentCommander
    - communicationsLiaison
    type: string
    x-enum-varnames:
    - IncidentCommander
    - CommunicationsLiaison
  incident.Service:
    properties:
      serviceId:
        type: string
      serviceName:
        type: string
    type: object
  incident.Severity:
    enum:
    - SEV1
    - SEV2
    - SEV3
    - SEV4
    type: string
    x-enum-varnames:
    - Severity1
    - Severity2
    - Severity3
    - Severity4
  incident.Slack:
    properties:
      channelDescription:
        type: string
      channelId:
        type: string
      channelName:
        type: string
      channelReference:
        type: string
      channelStatus:
        $ref: '#/definitions/incident.ChannelStatus'
    type: object
  incident.Stage:
    properties:
      createdAt:
        type: string
      duration:
        type: integer
      stageId:
        type: string
      type:
        $ref: '#/definitions/incident.Status'
      updatedAt:
        type: string
      userDetails:
        $ref: '#/definitions/utils.UserDetails'
    type: object
  incident.Status:
    enum:
    - started
    - detected
    - acknowledged
    - identified
    - mitigated
    - investigating
    - postmortemStarted
    - PostmortemCompleted
    - resolved
    type: string
    x-enum-varnames:
    - Started
    - Detected
    - Acknowledged
    - Identified
    - Mitigated
    - Investigating
    - PostmortemStarted
    - PostmortemCompleted
    - Resolved
  incident.Type:
    enum:
    - availability
    - latency
    - security
    - other
    type: string
    x-enum-varnames:
    - Availability
    - Latency
    - Security
    - Other
...
  utils.DefaultResponseDTO:
    properties:
      correlationId:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  utils.Source:
    enum:
    - slack
    type: string
    x-enum-varnames:
    - Slack
  utils.UserDetails:
    properties:
      email:
        type: string
      source:
        $ref: '#/definitions/utils.Source'
      userId:
        type: string
      userName:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: RespondNow
  description: RespondNow Server APIs.
  termsOfService: https://www.harness.io/legal/subscription-terms
  title: RespondNow Server
  version: "1.0"
paths:
...
  /incident/create:
    post:
      consumes:
      - application/json
      description: Create an incident
      operationId: CreateIncident
      parameters:
      - description: Create an incident
        in: body
        name: incident"
        required: true
        schema:
          $ref: '#/definitions/incident.CreateRequest'
      - description: accountIdentifier
        in: query
        name: accountIdentifier
        required: true
        type: string
      - description: orgIdentifier
        in: query
        name: orgIdentifier
        type: string
      - description: projectIdentifier
        in: query
        name: projectIdentifier
        type: string
      - description: correlationId
        in: query
        name: correlationId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/incident.CreateResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
      security:
      - ApiKeyAuth: []
      summary: Create an incident
      tags:
      - Incident
  /incident/list:
    get:
      consumes:
      - application/json
      description: List incidents
      operationId: ListIncidents
      parameters:
      - description: accountIdentifier
        in: query
        name: accountIdentifier
        required: true
        type: string
      - description: orgIdentifier
        in: query
        name: orgIdentifier
        type: string
      - description: projectIdentifier
        in: query
        name: projectIdentifier
        type: string
      - description: type
        enum:
        - availability
        - latency
        - security
        - other
        in: query
        name: type
        type: string
      - description: severity
        enum:
        - SEV1
        - SEV2
        - SEV3
        - SEV4
        in: query
        name: severity
        type: string
      - description: status
        enum:
        - started
        - detected
        - acknowledged
        - identified
        - mitigated
        - investigating
        - postmortemStarted
        - PostmortemCompleted
        - resolved
        in: query
        name: status
        type: string
      - description: active
        in: query
        name: active
        type: boolean
      - description: incidentChannelType
        enum:
        - slack
        in: query
        name: incidentChannelType
        type: string
      - description: search
        in: query
        name: search
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: correlationId
        in: query
        name: correlationId
        type: string
      - description: all
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/incident.ListResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.DefaultResponseDTO'
      summary: List incidents
      tags:
      - Incident Management
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"